Add files via upload
This script will grab a list of trading shares from the ASX website. It will then grab the latest JSON file for each share and upload the JSON to the S3 bucket for later use. This script will be run every 20 minutes during ASX trading hours, as automated by the EC2 OS the script will be hosted on.