Add files via upload
Script will now only download the company listing CSV from the ASX at 10:00AM (first run). During this process, it will generate a CSV containing only the ASX Company codes, and save this CSV locally. The Script from that point onward in the day will refer to the local CSV file for it's list of company codes, rather than download the ASX supplied CSV every time. 

The ASX CSV gets updated daily at midnight, so to redownload it every 20 minutes and not actually save it anywhere was a bit of a waste.