Add files via upload
This is an initial version of the user login server that will be always running on the EC2 instance. It will handle login requests when users try to log into their account on the application, and will handle the registration of new users.

This initial version is very buggy, and throws a lot of exceptions. IT IS A WORK I PROGRESS.