Add files via upload
This is a server module that will always be running on the EC2 instance. It will listen for calls from the front end user client, to handle login, and new registration requests. User info is stored privately in the S3 instance, and therefore the AWS key and secret keys are needed to access the data. This data has been redacted from the code.

New registrations create two files. One called creds/*emailHash*.rec inside the asx-user-store bucket, the other called data/*uniqueID*.obj. The .rec file contains thee hash of the user's password, and their unique ID number. The .obj file at the moment stores the a JSON entry representing the users profile data, but in later iterations, this file will be the byte stream of an object, rather than json in plain text.

When a login request is received, the server opens the file that matches the provided hash of the email address entered by the user on login, and compares the hash of the entered password, with the stored hash. If they match, the se