Added user transaction logging
Added user transaction logging
When the save function is invoked by the user application, the application should send the updated data string, as well as a transaction summary in the form of; {'Date':'04/04/2017', 'Time':'14:28', 'ASX_Code':'AMP', 'Qty':'+4', 'Price':'5.15'}.
Date and Time give the date and time at the time of the trade
ASX_Code refers to the 3 letter stock code on the ASX
Qty gives the quantities of the share acquired, positive being shares bought, negative being shares sold.
Price gives the price per share at the time of the trade.
This string is appended onto the end of the user's data file in S3 storage, each entry is separated by a new line feed. This data is part of the data returned to the user application when the user logs in, and will be used to display the users transaction history.