Created Score Updater
scoreUpdate.java is a small program that will be periodically invoked by the EC2 server. It will run on the 10min, 30min, and 50min mark between 10:00AM and 6:00PM.
It reads the shares owned by each player, and updates the players score based on the latest share value present on the S3 storage. This also updates the leaderboard accordingly.