Added stockHistory function
This function returns the value history of a particular stock code over a period of time. User's request a stock code, a start date, and an end date, and the server will return the different values at end of day between and including the given dates, for the given stock code.