Added sendMessage fcuntion
Allows users to send messages between each other
User program sends the hash of the senders email address, the hash of the intended recipients email, the message type, and the contents of the message.
Messages are saved in the individual users data folder, under mailbox. EG, data/100012/mailbox/*mailNo*.json
Mail is stored as JSON, with the attributes 'sender', 'type', and 'contents'.