moved creds to an abstract class
Moved the credentials to an abstract class called creds - won't be uploaded to github

Alive stocks now get added to an ArrayList in the manager called stockCodes. The codes in this list will be bubble sorted alphabetically, and this list will replace the list stored locally and in the S3 once all stocks have been downloaded.