Compacted code and added unread mail features
Moved some recurring code blocks into their own functions, such as posted data to the S3, reading data from the S3, and converting a given users email hash to their unique ID number.
Also implemented the ability for messages to be either "read" or "unread". This is done through a JSON tag; "Unread". Messages are automatically marked as unread when sent. When a user retrieves a message, it is automatically marked as read. Also added functionality for a user to mark a previously read message as unread.