Added sending funds between users funcitonality
User's can send funds to each other. When a user receives a funds transfer notification, they can accept the transfer, and accept it in full, or in part. Once a funds transfer has been accepted, the sender is notified, and told how much of the transfer was accepted, with any part not accepted being deposited back in their account.
   Coming Soon: Automated script to find funds transfers that haven't been actioned after 5 days and automatically put the money back in the senders account.