Create messageCheck.java
Message Check is a threaded script that runs after a user logs in, and terminates when the user logs out. Every minute, the thread queries the user's mailbox to check for any new messages or fund transfers. In the future, upon detecting a new message, this thread will call a function to alert the user in the UI